app/src/main/java/com/englishlearn/myapplication/data/TractateGroup.java
package com.englishlearn.myapplication.data;

import java.io.Serializable;

/**
 * Created by yanzl on 16-9-3.
 */
public class TractateGroup implements Serializable,Cloneable{

    //标记一对多关系
    private String __type;
    private String className;

    private String objectId;
    private boolean open; //是否公开
    private String name; //分组名称
    private User userId; //用户Id

    public void setPointer(){
        __type = "Pointer";
        className = "TractateGroup";
    }

    public String getObjectId() {
        return objectId;
    }

    public void setObjectId(String objectId) {
        this.objectId = objectId;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public User getUserId() {
        return userId;
    }

    public void setUserId(User userId) {
        this.userId = userId;
    }

    public boolean isOpen() {
        return open;
    }

    public void setOpen(boolean open) {
        this.open = open;
    }

    @Override
    public Object clone(){

        Object o = null;
        try {
            o = super.clone();
        } catch (CloneNotSupportedException e) {
            e.printStackTrace();
        }
        return o;
    }

    @Override
    public String toString() {
        return "TractateGroup{" +
                "objectId='" + objectId + '\'' +
                ", open=" + open +
                ", name='" + name + '\'' +
                ", userId=" + userId +
                '}';
    }
}
